refactor: move homepage components to candidate route
