--- web/src/routes/(marketing)/product-features.svelte
@@ -1,86 +1,0 @@
-<script lang="ts">
-	const features = [
-		{
-			icon: `
-  <path stroke-linecap="round" stroke-linejoin="round" d="M21.75 6.75v10.5a2.25 2.25 0 01-2.25 2.25h-15a2.25 2.25 0 01-2.25-2.25V6.75m19.5 0A2.25 2.25 0 0019.5 4.5h-15a2.25 2.25 0 00-2.25 2.25m19.5 0v.243a2.25 2.25 0 01-1.07 1.916l-7.5 4.615a2.25 2.25 0 01-2.36 0L3.32 8.91a2.25 2.25 0 01-1.07-1.916V6.75" />
-			`,
-			title: 'Free Your Inbox',
-			description:
-				'Keep your inbox free of unwanted inbound opportunities. Automated sequences are stopped in their tracks.'
-		},
-		{
-			icon: `
-<path stroke-linecap="round" stroke-linejoin="round" d="M21 21l-5.197-5.197m0 0A7.5 7.5 0 105.196 5.196a7.5 7.5 0 0010.607 10.607zM10.5 7.5v6m3-3h-6" />
-			`,
-			title: 'Miss Nothing',
-			description:
-				'SRC syncs with your inbox to keep track of all past, present, and future job opportunities.'
-		},
-		{
-			icon: `
-  <path stroke-linecap="round" stroke-linejoin="round" d="M11.25 4.5l7.5 7.5-7.5 7.5m-6-15l7.5 7.5-7.5 7.5" />
-			`,
-			title: 'Hit the Ground Running',
-			description:
-				'Once you are ready for a new opportunity, SRC will automatically re-engage previous recruiters to find out which ones are still active.'
-		},
-		{
-			icon: `
-  <path stroke-linecap="round" stroke-linejoin="round" d="M11.42 15.17L17.25 21A2.652 2.652 0 0021 17.25l-5.877-5.877M11.42 15.17l2.496-3.03c.317-.384.74-.626 1.208-.766M11.42 15.17l-4.655 5.653a2.548 2.548 0 11-3.586-3.586l6.837-5.63m5.108-.233c.55-.164 1.163-.188 1.743-.14a4.5 4.5 0 004.486-6.336l-3.276 3.277a3.004 3.004 0 01-2.25-2.25l3.276-3.276a4.5 4.5 0 00-6.336 4.486c.091 1.076-.071 2.264-.904 2.95l-.102.085m-1.745 1.437L5.909 7.5H4.5L2.25 3.75l1.5-1.5L7.5 4.5v1.409l4.26 4.26m-1.745 1.437l1.745-1.437m6.615 8.206L15.75 15.75M4.867 19.125h.008v.008h-.008v-.008z" />
-			`,
-			title: 'No New Tools',
-			description:
-				"SRC isn't another just another job search platform. It lets you get the most out of your existing tools and inbound opportunities"
-		},
-		{
-			icon: `
-  <path stroke-linecap="round" stroke-linejoin="round" d="M12 3v17.25m0 0c-1.472 0-2.882.265-4.185.75M12 20.25c1.472 0 2.882.265 4.185.75M18.75 4.97A48.416 48.416 0 0012 4.5c-2.291 0-4.545.16-6.75.47m13.5 0c1.01.143 2.01.317 3 .52m-3-.52l2.62 10.726c.122.499-.106 1.028-.589 1.202a5.988 5.988 0 01-2.031.352 5.988 5.988 0 01-2.031-.352c-.483-.174-.711-.703-.59-1.202L18.75 4.971zm-16.5.52c.99-.203 1.99-.377 3-.52m0 0l2.62 10.726c.122.499-.106 1.028-.589 1.202a5.989 5.989 0 01-2.031.352 5.989 5.989 0 01-2.031-.352c-.483-.174-.711-.703-.59-1.202L5.25 4.971z" />
-			`,
-			title: 'Level the Playing Field',
-			description:
-				'Companies have an arsenal of tools to help them find the best candidates. SRC gives you the same tools to help you find the best opportunities.'
-		},
-		{
-			icon: `
-  <path stroke-linecap="round" stroke-linejoin="round" d="M3.75 6.75h16.5M3.75 12H12m-8.25 5.25h16.5" />
-			`,
-			title: 'One Format, Endless Opportunities',
-			description:
-				'SRC creates a single, easy-to-read format for all your job opportunities. No more searching through emails or spreadsheets.'
-		}
-	];
-</script>
-
-<section
-	id="product"
-	class="flex min-h-screen flex-col items-center bg-slate-900 py-20 text-blue-300"
->
-	<h2 class="max-w-4xl text-center text-4xl font-bold sm:text-5xl">
-		Your Personal Recruiting Assistant
-	</h2>
-	<h3 class="mt-8 mb-16 max-w-2xl px-4 text-center text-slate-100">
-		SRC puts candidates first in the recruiting process, respecting your time and privacy.
-	</h3>
-	<div class="flex w-2/3 flex-wrap justify-between">
-		{#each features as feature}
-			<div
-				class="mb-12 flex max-w-xs flex-col items-center space-y-4 text-center sm:items-start sm:text-left"
-			>
-				<svg
-					xmlns="http://www.w3.org/2000/svg"
-					class="text-secondary-focus h-12 w-12"
-					fill="none"
-					viewBox="0 0 24 24"
-					stroke="currentColor"
-					stroke-width="1.5"
-				>
-					{@html feature.icon}
-				</svg>
-				<p class="text-secondary-focus text-xl font-semibold">{feature.title}</p>
-				<p class="text-slate-100">
-					{feature.description}
-				</p>
-			</div>
-		{/each}
-	</div>
-</section>
